Link to the latest MSVC 2015/2017/2019 redistributable
MSVC 2015, 2017 and 2019 all share the same redistributable, so
just link to the latest version of that redistributable.

The installer will also download and install the latest version of the
redistributable. The check to see if the redist is already installed now
compares redist version numbers. The installer would previously only
check if the version it was about to download was installed, which meant
that having a newer version installed would cause the redist to be
downloaded and installed pointlessly. The version that is checked for is
the same 2017 redist version that was previously linked to, because
there's apparently no way to tell what the actual minimum required
version is.